pyproject.toml
[project]
name = "phenotastic"
version = "0.3.0"
description = "3D plant phenotyping."
readme = "README.rst"
requires-python = ">=3.12,<4.0.0"
license = { text = "GNU General Public License v3" }
authors = [
    { name = "Henrik Ahl", email = "henrikaahl@gmail.com" }
]
keywords = ["phenotastic", "3D", "phenotyping", "segmentation"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.15",
]
dependencies = [
    "imgmisc>=0.1.17",
    "mahotas>=1.4.13",
    "numpy>=2.4.2",
    "pyvista>=0.46.5",
    "vtk>=9.5.2",
    "pystackreg>=0.2.8",
    "pandas>=3.0.0",
    "scikit-image>=0.26.0",
    "psf>=2026.1.18",
    "scipy>=1.17.0",
    "loguru>=0.7.3",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
phenotastic = "phenotastic.cli:main"

[project.urls]
homepage = "https://github.com/supersubscript/phenotastic"
documentation = "https://phenotastic.readthedocs.io"
repository = "https://github.com/supersubscript/phenotastic"

[project.optional-dependencies]
docs = [
    "Sphinx>=5.1.1",
]

[dependency-groups]
dev = [
  "ipython>=8.26.0,<9.0.0",
  "mypy>=1.10.1,<2.0.0",
  "pre-commit>=3.7.1,<4.0.0",
  "pytest>=8.2.2,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.7.2,<1.0.0",
  "seaborn>=0.13.2",
  "pyqt5>=5.15.11",
  "ipykernel>=6.29.5",
  "types-PyYAML>=6.0",
]
release = ["python-semantic-release>=9.19.1,<10.0.0"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/phenotastic"]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.mypy]
allow_redefinition = true
cache_dir = ".mypy_cache"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
explicit_package_bases = true
follow_imports = "skip"
ignore_missing_imports = true
incremental = false
namespace_packages = true
python_version = "3.12"
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false

[tool.ruff]
fix = true
extend-include = ["*.ipynb"]
line-length = 120
show-fixes = true
target-version = "py312"

[tool.ruff.lint]
select = [
    "I",    # isort
    "F",    # Pyflakes
    "E",    # pycodestyle (errors)
    "W",    # pycodestyle (warnings)
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "PT",   # flake8-pytest-style
    "UP",   # pyupgrade
    "TID",  # flake8-tidy-imports
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "FLY",  # flynt
    "TRY",  # tryceratops
    "NPY",  # NumPy-specific rules
    "C901", # Cognitive complexity
]
ignore = [
    "B023",   # allow using local variable in for-loops
    "NPY002", # allow legacy np.random usage
    "TRY003", # allow long messages in exceptions
    "A005",   # allow shadowing module names
]

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.coverage.run]
omit = []

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'def __repr__',
    'if self.debug:',
    'if settings.DEBUG',
    'raise AssertionError',
    'raise NotImplementedError',
    'if 0:',
    'if __name__ == "__main__":',
    'def main',
]
show_missing = true

[tool.semantic_release]
version_variables = ["phenotastic/__init__.py:__version__"]
version_toml = ["pyproject.toml:project.version"]
upload_to_PyPI = false
upload_to_release = true
build_command = "uv lock && uv build"
commit_message = "build(release): update version to {version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"
assets = ["uv.lock"]

[tool.semantic_release.branches.main]
match = "main"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "feat",
    "fix",
    "perf",
    "style",
    "refactor",
    "test",
    "wip",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false
token = { env = "GITHUB_TOKEN" }

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
